requirements.txt
aiofiles==23.2.1
aiosqlite==0.20.0
alembic==1.13.1
black==24.4.2
click==8.1.7
greenlet==3.0.3
Mako==1.3.5
MarkupSafe==2.1.5
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
python-dotenv==1.0.1
SQLAlchemy==2.0.30
tomli==2.0.1
typing_extensions==4.12.2
